% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial_utils.R
\name{add_city_codes}
\alias{add_city_codes}
\title{Add city codes to a data frame}
\usage{
add_city_codes(adf)
}
\arguments{
\item{adf}{}
}
\value{
tibble
}
\description{
Add city codes to a data frame
}
