% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial_utils.R
\name{add_country_codes}
\alias{add_country_codes}
\title{Add country codes to a data frame}
\usage{
add_country_codes(adf)
}
\arguments{
\item{adf}{}
}
\value{
tibble
}
\description{
Add country codes to a data frame
}
