% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial_utils.R
\name{add_subregion_codes}
\alias{add_subregion_codes}
\title{Add subregion codes to a data frame}
\usage{
add_subregion_codes(adf)
}
\arguments{
\item{adf}{}
}
\value{
tibble
}
\description{
Add subregion codes to a data frame
}
