% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_covid19mobility.R
\name{get_info_covid19mobility}
\alias{get_info_covid19mobility}
\title{Get information about the datasets provided by covid19mobility}
\usage{
get_info_covid19mobility()
}
\value{
a tibble of information about the datasets in this package
}
\description{
Returns information about the datasets in this package for covid19R harvesting
}
\examples{
\dontrun{

# get the dataset info from this package
get_info_covid19mobility()
}

}
