% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial_utils.R
\name{location_code_merge_pivot}
\alias{location_code_merge_pivot}
\title{Merge and pivot data in a standardized way}
\usage{
location_code_merge_pivot(codes, adf, name_col, code_col, code_name = code_col)
}
\arguments{
\item{codes}{tibble with a column containing standardized codes}

\item{adf}{tibble of data}

\item{name_col}{what column in the codes tibble contains the name matching info}

\item{code_col}{what column in the codes tibble contains the standardized codes}

\item{code_name}{what is the name of the code standard - defaults to the same as code_col}
}
\value{

}
\description{
Merge and pivot data in a standardized way
}
