% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mobility.R
\name{mod_mobility_ui}
\alias{mod_mobility_ui}
\title{mod_mobility.R}
\usage{
mod_mobility_ui(id)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
