% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_covid19mobility_apple.R
\name{refresh_covid19mobility_apple_city}
\alias{refresh_covid19mobility_apple_city}
\title{Refresh The Apple Covid-19 Mobility Data for Cities}
\usage{
refresh_covid19mobility_apple_city()
}
\value{
Returns a tibble that meets the Covid19R Project tidy data standard
}
\description{
Pulls in the CSV of the Apple Mobility Data, filters to cities, and reshapes it
}
\examples{
\donttest{

mob <- refresh_covid19mobility_apple_city()

head(mob)
}

}
\references{
\url{https://www.apple.com/covid19/mobility}
}
