% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_covid19mobility_apple.R
\name{refresh_covid19mobility_apple_subregion}
\alias{refresh_covid19mobility_apple_subregion}
\title{Refresh The Apple Covid-19 Mobility Data for Subregions}
\usage{
refresh_covid19mobility_apple_subregion()
}
\value{
Returns a tibble that meets the Covid19R Project tidy data standard
}
\description{
Pulls in the CSV of the Apple Mobility Data, filters to subregions, and reshapes it
}
\examples{
\donttest{

mob <- refresh_covid19mobility_apple_subregion()

head(mob)
}

}
\references{
\url{https://www.apple.com/covid19/mobility}
}
