% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_covid19mobility.R
\name{refresh_covid19mobility_country}
\alias{refresh_covid19mobility_country}
\title{Refresh The Apple Covid-19 Mobility Data for Countries}
\usage{
refresh_covid19mobility_country()
}
\value{
Returns a tibble that meets the Covid19R Project tidy data standard
}
\description{
Pulls in the CSV of the Apple Mobility Data, filters to country, and reshapes it
}
\examples{
\dontrun{

mob <- refresh_covid19mobility_country()

head(mob)
}

}
\references{
\url{https://www.apple.com/covid19/mobility}
}
