% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid19mobility_google_country_demo}
\alias{covid19mobility_google_country_demo}
\title{Demo Data of the Apple Covid-19 Mobility Data for Countries}
\format{
A data frame with 83160 rows and 7 variables:
\describe{
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
covid19mobility_google_country_demo
}
\description{
Demo Data of the Apple Covid-19 Mobility Data for Countries
}
\keyword{datasets}
